Install MySQL client packages (#1525)
This change will install MySQL client packages in order to support Slurm
accounting.

In order to validate the installation, I've added a recipe to the main
cookbook for image validation which will call to a validation recipe
in the aws-parallelcluster-tests cookbook. This recipe will execute
the MySQL package validation to make sure that the installed packages
come from the expected repositories dependent on platform and
architecture.

This validation requires a forthcoming change to the ImageBuilder test
component template in the aws-parallelcluster repository in order to
actually execute.