Change default NFS thread number to have better performances
I tested to launch a job with 1000 nodes on a c5.12xlarge (24 NFS threads) and
I noticed the `nfsd: too many open connections, consider increasing the number of threads` message.

We should use at the very least one daemon per processor,
but four to eight per processor is considered a better rule of thumb.
We're capping the value to 256 to avoid having too many threads.

With this patch I'm also exposing the `node['cluster']['nfs']['threads']` parameter in the `default.rb`
file to permit to overwrite it from `DevSettings` if needed.

Signed-off-by: Enrico Usai <usai@amazon.com>