Use source_ami in place of source_ami_filter for CentOS7
This is required because CentOS 7 AMI for x86_64 arch has been deprecated
and packer does not permit to retrieve it.

In this way the custom_ami_id is retrieved upstream (by using --include-deprecated flag)
and passed as input as CUSTOM_AMI_ID environment variable.

This change doesn't impact CentOS 7 arm because in that case we're
using the custom-centos7 builder.

Signed-off-by: Enrico Usai <usai@amazon.com>