surface head node bootstrap errors - store error messages from custom scripts
During the bootstrap of the head node, if any of custom scripts (OnNodeStart or OnNodeConfigured) fails, a generic error message will be sent to CloudFormation stack without providing specific reasons of the failure. Here we write error messages into a dedicated file, which will be extracted and sent to CloudFormation stack. Also updated the error file name from headnode_bootstrap_error_msg to bootstrap_error_msg, since this file will be used in both the head node and compute nodes.