Move ulimit configuration in users.rb and add tests for users.rb
With this commit we're adding a new suite to test the users control.
Every suite can be considered a unit test that will execute a specific run list
and specific controls during the test phase.

Remove suites from kitchen dokken tests to execute all of them in the same GitHub action.

Signed-off-by: Edoardo Antonini <eantonin@amazon.com>
Signed-off-by: Enrico Usai <usai@amazon.com>