Merge Wip/headnode bootstrap error branch to develop (#1688)
* rename the file to store error messages from both chef and custom scripts

* surface head node bootstrap errors - store error messages from custom scripts

During the bootstrap of the head node, if any of custom scripts (OnNodeStart or OnNodeConfigured) fails, a generic error message will be sent to CloudFormation stack without providing specific reasons of the failure. Here we write error messages into a dedicated file, which will be extracted and sent to CloudFormation stack. Also updated the error file name from headnode_bootstrap_error_msg to bootstrap_error_msg, since this file will be used in both the head node and compute nodes.

* create chef handler and apply to recipes

* rename error file

* add a check to avoid conflict with BYOS error handling

* refactor the code a little bit

* rename error file

* address reviewer comments

* correct year in license

* resolve Ruby check failure

* replace preinstall or postinstall with OnNodeStart or OnNodeConfigured in error messages

* add function to trim url when it is longer than url length limit

* revise the chef handler to detect mounting issue otherwise extract recipe information

* revise the logic to cover EFS and FSX mount errors

* clean up handler script and move to libraries folder

* fix Ruby Cookstyle offenses

* update chef handler according to review comments

Co-authored-by: Delong Meng <delmeng@amazon.com>
Co-authored-by: delongmeng <51895412+delongmeng@users.noreply.github.com>
Signed-off-by: Enrico Usai <usai@amazon.com>