Define new recipe to automatically import recipes and resources deps
This new recipe will permit to define tests requiring multiple dependencies
in term of resources and recipes.

Intead of using the run_list attribute to define the list of recipes
we are now using the dependencies attribute, on which it's possible to define
a list of deps, identified by the recipe: or resource: prefix.

Adapted existing tests accordingly.

Signed-off-by: Enrico Usai <usai@amazon.com>