Add test for ephemeral drives setup
The setup-ephemeral-drives.sh script requires the cfn_ephemeral_dir variable
in the cfnconfig file, I added the cfnconfig_mock.rb to create the required file.

Ephemeral drives test verifies the /scratch directory (from cfnconfig) is correctly mounted
in an instance type m5d.xlarge with 1 instance store.

lvm2 is already installed as part of the install_packages but it is required
to have the ephemeral_drives.rb recipe self contained.
The name of the package is the same in all the OSes.

Signed-off-by: Enrico Usai <usai@amazon.com>