Use start of string anchor regex for dir path ref
Previous to this commit, the pattern sent to `delete_lines` is an
(interpolated) string followed by a space, such as: `/nfs `, operating
on /etc/exports. Without the start of string anchor regex (`^`), the
string will match a different line that shares includes the string and
whitespace. Example:

`/shared `

will match:

`/shared ` (good)
`/opt/parallelcluster/shared ` (bad!)

Including a `^` forces the string to match from the beginning of the
line.