Publish JSON event logs to CloudWatch
The change allows for logs files to be generated
without requiring a timestamp on each record. This
is to support structured JSON logging. CloudWatch
will not parse the log event as JSON if a
timestamp precedes the JSON structure in the
log message.

This change also configures CloudWatch to publish 
new JSON events to log streams in the cluster log group.

Signed-off-by: David Pratt <davprat@amazon.com>