Create new resource to manage system authentication
The system_authentication resource is used at install time to install SSSD and related tools
and at config time to configure NSS, PAM to use SSSD for system authentication and identity information.

Moved SSSD packages from default attributes to system_authentication.

The authconfig tool used on centos7 and alinux2 is now in compatibility mode for rhel8
so it's required to migrate to authselect.

Added Inspec tests to verify installation and configuration phases (skipped on docker redhat).

Note: I'm missing verification test for debian systems.
pam-auth-update command that is used to configure the system cannot be used to verify the changes
because it's an interactive command and I wasn't able to find differences in related files.

Signed-off-by: Enrico Usai <usai@amazon.com>