Allow generation of empty custom Slurm setting include file
Always generate a custom Slurm setting include file, which will be
empty if no custom Slurm settings are used.

This avoids additional logic to delete the custom Slurm settings
include files if the custom Slurm settings are removed via a pcluster
update-cluster operation.

Add unit test to verify the generation of an empty include file if
CustomSlurmSettings is not used.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>