Add node Health Check Manager
The Health Check Manager execution is triggered by a Slurm prolog script and executes the health checks sequentially (today only GPU check). It logs the checks output and error to a new dedicated log and exit with the sum of the exit codes of the checks.

Health Check Manager code and Health Check scripts are shared from head node, so to be better maintainable.

Slurm prolog and epilog configurations are set to target a folder.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>