Handle Kitchen dependencies
Kitchen resolves relative dependencies based on the directory from which it was launched. Therefore, dependencies that are resolved correctly when launched from a sub-cookbook will not work when launched from the main cookbook.

To avoid changing the codebase, dependencies are copied to a temporary folder, where the paths are adjusted based on the launching directory.