Add network service kitchen tests
The network service has two actions, restart and reload.  This commit adds a control to check if the
service is running with a switch for different operating systems.  There are two additional controls
that verify the service is running after the two different actions defined for the service.

The reload action cannot be performed on docker, so that is skipped when virtualized