Restructure CloudWatch Inspec test to avoid failures on Ubuntu20
Inspec test executed in the validate phase after the build of the image is failing on the
gpg --list-keys command because the exit code is 2.
Anyway the cloudwatch keys are in the output.

This started to fail when we added the tag:install to the cloudwatch_installation_files control
but it does not fail when executed as a singular inspec test.

For this reason I'm currently restructuring the Inspec code to avoid to check
the exit code in the Ubuntu20 case.

Tested executing Inspec tests for all OSes on both docker and EC2

Signed-off-by: Enrico Usai <usai@amazon.com>