Explicitly download all the Intel HPC dependencies
The installation was failing with errors like:
```
STDERR: Error downloading packages:
  libXcursor-1.1.15-1.el7.x86_64: Caching enabled but no local cache of /var/cache/yum/x86_64/7/base/packages/libXcursor-1.1.15-1.el7.x86_64.rpm from base/7/x86_64
  libXt-1.1.5-3.el7.x86_64: Caching enabled but no local cache of /var/cache/yum/x86_64/7/base/packages/libXt-1.1.5-3.el7.x86_64.rpm from base/7/x86_64
...
```
This happens because we're installing the packages with `--cacheonly` flag that installs
entirely from system cache, so packages must be explicitly downloaded and available in the cache.

They were probably installed as part of Gnome package group or other packages installed as
part of recipes executed before the intel.rb. With this change we're sure the intel recipe is self-contained.

Signed-off-by: Enrico Usai <usai@amazon.com>