Do not use DCV file exist to verify if DCV is configured or not
ParallelCluster recipes are creating the DCV configuration file from a template file
at instance bootstrap time if DCV is enabled, but the file is created at DCV setup.

Check the file exists is not enough, we need to verify if DCV is currently enabled.

This issue was causing the addition of dcv authenticator daemon in the supervisord template
even if DCV was not enable and this was causing failures in the recipe execution
because dcv_port was set to NONE by CDK generation so it was not possible to convert it to Integer.

### Tests

Created a unit test to verify supervisord template creation.

### References

Issue introduced with: 9467c928784703d94ca9010c08215192adaeab65

Signed-off-by: Enrico Usai <usai@amazon.com>