Add Manage EBS kitchen tests to cover the manage_ebs resource
This commit leverages the volume resource to perform ebs filesystem management tasks and adds
inspec kitchen tests and chefspec unit tests to cover those resources.  The kitchen tests currently
require that EBS volumes are manually created in advance and are referenced in the test suite config.
None of these tests can be run in docker since networking info is not available and ebs volumes are not
available.