Add kitchen test support for EFS mount and unmount actions for all OSes, including Ubuntu 22.
EBS kitchen tests existed for the install utils action, but not yet for mount and unmount.
Updated the mount and unmount actions to check for a nil array on optional properties.
Similar to EBS, the tests are executed with manually created EFS filesystems, so not CI/CD.
Remove unnecessary stunnel describe block.
The following controls were added:
- efs_mounted
- efs_unmounted

The following controls were tested manually:
- efs_install_utils on ec2 and docker (only rhel8 no-op)
- efs_mount on ec2 and docker (all no-op)
- efs_unmount on ec2 and docker (all no-op)