Introduce a new approach to share info from lifecycle hooks to kitchen environment
`kitchen.global.yml` configures lifecycle hooks to run potentially pre and post every operation.
`kitchen.run-hook.sh` verifies if the hook script exists under `<cookbook>/test/hooks/<install|config>/<suite>`.
If the script exists, it is called.

`kitchen.global.yml` uses kitchen environment, which is configured in `test/environments/kitchen.rb`.

In order to share values with kitchen recipes running remotely, attribute values must be set in this file.

For every value to pass and for every OS, you need add a line to this file:
```
'<suite_name>-<variable_name>/<platform>' => 'placeholder'
```

For instance: `'ebs_mount-vol_array/alinux2' => 'placeholder'`.

In your hook scripts, use sed to replace the placeholder with the actual value to share with the remote host.
Also, read the value from the same file to retrieve it locally,
for instance to be able to delete a resource during the destroy phase.

It creates an EBS volume using post_create hook,
shares it with hooks recipe which logs the value, and destroys it in pre_destroy hook.

If you use the "FROM_HOOK" keyword in the suite definition,
the `test_resource.rb` will automatically replace the placeholder for you in the environment file.

### Tests

* `./kitchen.ec2.sh environment-install test ebs-mount-alinux2`

```
[2023-06-09T08:09:49+00:00] INFO: FROM_HOOK found in resource properties. Original descriptor is: manage_ebs:mount {"shared_dir_array" : ["shared_dir"], "vol_array" : "FROM_HOOK-ebs_mount-vol_array"}
[2023-06-09T08:09:49+00:00] INFO: Hook key ebs_mount-vol_array/centos7 found in the environment. Replacing FROM_HOOK value with: ["vol-0ac6a0534c431319f"]
```

Signed-off-by: Enrico Usai <usai@amazon.com>