Add support for FSX on Ubuntu 22
Added controls for lustre mount and unmount.  The filesystem creation is not automated
yet using lifecycle hooks.  It leverages the test/environments/kitchen.rb file but with
manually added variable declarations.  Lustre requires 6 properties for mounting and
the environment file requires 6 entries for each property per operating system.  In order
to avoid the explosion of properties, the lifecycles hook mechanism will require lock
files and parallelism controls in order to share FSX filesystems with multiple simultaneous tests.
This effort will be completed in a future PR.

Completed manual tests for
-lustre:mount
-lustre:unmount