Fix bug in disable_ptrace action to allow sysctl to disable ptrace. (#2291)
Previously the setting for Ubuntu systems was not getting used due to a misalignment between the cookbook and cli.
On the cli, node['cluster']['enable_efa'] was set to 'efa' in the dna.json file, but the cookbook assumed the
parameter was set to 'compute'.  This causes the disable_ptrace action if clause to be skipped.
This page describes the use of sysctl https://manpages.ubuntu.com/manpages/bionic/man5/sysctl.conf.5.html
This page describes the function of the ptrace_scope parameter https://www.kernel.org/doc/Documentation/security/Yama.txt
Chef actually generates a config file in /etc/sysctl.d/ for the parameter that is changed so it should persist.
https://docs.chef.io/resources/sysctl/

Extended the efa_debian_system_settings_configured kitchen test to include the file addition but spec tests remain the same.