Replace nvidia-persistenced with parallelcluster_nvidia service
`parallelcluster_nvidia` service ensures the creation of the block devices `/dev/nvidia0`
and it is needed by the `slurmd` service.

`parallelcluster_nvidia` starts the `nvidia-persistenced` or runs `nvidia-smi`
to avoid race condition with other services and avoids conflicts when using DLAMI with a gpu instance.

### Tests
* Modified ChefSpec to verify new changes.

### References
Backport of: https://github.com/aws/aws-parallelcluster-cookbook/pull/2341

Signed-off-by: Enrico Usai <usai@amazon.com>