Create new mechanism to execute OS specific hooks
The kitchen tests will automatically execute hooks from kitchen/hooks folder
before executing the hooks for the specific test.

For example we can have a `kitchen/hooks/alinux2/pre_converge.sh` with the following content,
to automatically install some packages in the instance right after the startup, before Chef installation:

```
#!/bin/bash

if [ -n "${KITCHEN_INSTANCE_HOSTNAME}" ]; then
    export KITCHEN_EC2_USER='ec2-user'
    echo "Install libxcrypt-compat package by using SSH key: ${KITCHEN_SSH_KEY_PATH}"

    KITCHEN_EC2_INSTANCE_ID=$(ssh -o StrictHostKeyChecking=no -i "${KITCHEN_SSH_KEY_PATH}" \
      "${KITCHEN_EC2_USER}@${KITCHEN_INSTANCE_HOSTNAME}" 'sudo yum install -y some-package')
fi
```

This is useful for example when Chef cannot be installed in the AMI because due to a missing or
conflicting packages.

Signed-off-by: Enrico Usai <usai@amazon.com>