Move Intel HPC configuration after setup of shared folder
The config step of the Intel HPC installer installs rpms from `/opt/intel/rpms/`.
This folder must be mounted before executing intel hpc installation steps.

Without this patch the installation works for the head node but
fails for the compute nodes because unable to see the rpms in the shared folder.

Signed-off-by: Enrico Usai <usai@amazon.com>