Move ulimit check to install phase and switch user to reflect changes to the config
Until now ulimit -Sn was used in the config step to confirm the changes in the install
step are reflected in the running OS.  This is no longer working in Ubuntu 22 because
for whatever reason, the session that is created for Ubuntu 22 doesn't have the
changes reflected, even though the files are in place.  After launching an instance
from the AMI however, the changes are there.  To improve this test, we need to move
it to the install step and check it immediately by switching users and forcing the
loading of the limits.