Remove useless pre_verify step to fix inspec file path
Kitchen resolves relative dependencies based on the directory from which it was launched.
In the past, the structure of the repo was composed by a main cookbook and multiple sub-cookbooks.
Therefore, dependencies that were resolved correctly when launched from a sub-cookbook were not working when launched from the main cookbook.

To avoid changing the codebase, we were copying dependencies to a temporary folder,
where the paths were adjusted based on the launching directory.

This is no longer required, since we totally refactored the repo structure,
and it's no longer required to jump into the cookbook to kitchen test it.

For sure the sed is useless since the `path: ../aws-parallelcluster` doesn't exist in the inspec files.

### References
* https://github.com/aws/aws-parallelcluster-cookbook/commit/d12e41286d752d1d07a5fbab16b1d2436540c353
* https://github.com/aws/aws-parallelcluster-cookbook/commit/5aabe07639d2964279a52e199c46664d78128516

Signed-off-by: Enrico Usai <usai@amazon.com>