Use custom_node_package and custom_awsbatchcli_package only when running system tests
These packages are not required in CI/CD because only config recipes are executed there,
then they are not required by local ec2 or docker tests, because locally we execute specific
tests/controls that should have all the required parameters.

With this change we're also avoiding to install always version 3.4.1 of the node
and use instead a package from develop.

### Tests

* Created new test for custom awsbatch cli package installation
* Improved existing test for custom node package installation to check if the package is available on pypi

Executed:
* `bash kitchen.ec2.sh computefleet-install test custom-parallelcluster-node-rocky8`
* `bash kitchen.ec2.sh awsbatch-install test custom-awsbatchcli-package-rocky8`

Signed-off-by: Enrico Usai <usai@amazon.com>