Add support for EC2 Capacity Blocks for ML
Modified pcluster_fleet_config_generator to save Capacity Reservation id and
Capacity Type info in the config-fleet.json file.

These informations will be used by the node logic to query Capacity Block
reservation and verify if they are active or not.

Fleet config now contains, for every compute resource:
- CapacityType (previously was available on create-fleet case only, now it has a default "ondemand" value)
- CapacityReservationId (taken from compute resource or from the queue)

Previous parameters are taken from compute resource config, or from the queue config.
CapacityType and AllocationStrategy are not yet supported at compute resource level
from the CLI, but this code is ready to read them.

AllocationStrategy is now set in the generated config only if it has a value
because for capacity-block this is not supported.

Note: CapacityReservationResourceGroup is not taken into consideration.
Semantic validation of the configuration parameter is performed at CLI level.

### Tests
Extended unit tests to cover generation of the new info

Signed-off-by: Enrico Usai <usai@amazon.com>