[Develop] Disable Spack until the feature is complete (#2530)
Spack was initially added with several optimizations for a range of applications in mind.  This resulted
in several packages.yaml files for different OS families and make opinionated choices about what to
install and how.  This could conflict with some power users' use of the installation.  Because of this
the final released implementation is on hold until consensus can be reached for what to release.