Fix bug introduced with DatabaseName functionality
When Slurm Accounting is used but DatabaseName is not defined in the
input config, it is rendered as `null` in the config on the cluster,
which is translated to `none` in Jinja2.

This means that the value passes the `is defined` test in Jinja2.

Fix the if condition in Jinja2 to check if `DatabaseName` is not `none`
before trying to set it in `StorageLoc` in the slurmdbd configuration.

Fix the unit test to have `DatabaseName` equal to `null` in the case
when Slurm Accounting is used with default database name.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>