Upgrade nfs third-party cookbook to version nfs-5.1.2 (from nfs-5.0.0)
Version 5.0.1 introduced support for Ubuntu22, anyway we weren't able to upgrade
the package because nfs-utils `1:2.6.1-1ubuntu1.2` wasn't correctly supported.

I created a patch for nfs cookbook to fix this incompatibility issue with nfs-utils.
This has been released, so we can now upgrade to 5.1.x and remove the `ubuntu.22.json` file.

### References
* https://github.com/sous-chefs/nfs/releases
* https://github.com/sous-chefs/nfs/pull/142
* https://bugs.launchpad.net/ubuntu/+source/nfs-utils/+bug/1971935

Signed-off-by: Enrico Usai <usai@amazon.com>