Remove useless mechanism to retrieve OS minor version from RHEL and Rocky
This mechanism was used on CentOS 7, to retrieve the OS minor version from the kernel version
and use it to download the right version of the Lustre client.

This mechanism is not required for RHEL and Rocky because the minor version
automatically changes when updating the kernel. This is reflected in the `/etc/os-release` file too.

This change will permit us to automatically support new minor versions when they will be available,
without the need to update the internal mapping.

Signed-off-by: Enrico Usai <usai@amazon.com>