Install kernel-devel for Rocky Linux from vault
Previous releases are moved into a vault area once a new minor release version is available for at least a week.
This means we cannot use `--releasever` to install 8.8 version because it is not more available in the repo.

On RHEL8 instead all versions of the packages will remain available.

With this patch we're downloading this specific package from vault and installing it.

## Tests

Manual execution of the new steps.

## References

* https://wiki.rockylinux.org/rocky/repo/#notes-on-devel

Signed-off-by: Enrico Usai <usai@amazon.com>