Remove aws-parallelcluster-shared/test dependency from other cookbooks
A Chef InSpec profile can bring in the controls and custom resources from another Chef InSpec profile.
This dependency is not required because we don't need to execute controls or custom resources from aws-parallelcluster-shared cookbook.

Removing this dependency we should solve the concurrency issue causing random failures with the error:
```
[2023-12-14T23:25:37.069Z] >>>>>>     Failed to complete #verify action: [No such file or directory @ rb_file_s_stat - /var/jenkins_home/jobs/kitchen_tests_3/workspace/pcluster_cookbook/cookbooks/aws-parallelcluster-shared/test/inspec.json] on slurm-config-compute-fleet-arm64-t4gxlarge-alinux2
```

### References
* https://docs.chef.io/inspec/profiles/depends/

Signed-off-by: Enrico Usai <usai@amazon.com>