Remove mysql_repo resource
The package provided by mysql to configure mysql-repos ships a key
that will eventually expire and cause errors during an apt-update
since the verification of the repositories will fail.
We are going to remove it since the benefit of having it configured
in the AMI weight less than the negative aspects.

Signed-off-by: Nicola Sirena <nssirena@amazon.com>