[DFSM] Fix race conditions when checking if a live update is supported by a storage change.
In particular, we moved the check that determines if a storage change can be applied with a live update, from a lambda block to specific only_if blocks. This is necessary to let Chef evaluate it when it gets executed, rather than at the beginning of the converge phase.

Also, removed unnecessary state 'SKIPPED_LIVE_UPDATE'
and covered DFSM critical functionalities with spec tests.

Signed-off-by: Giacomo Marciani <mgiacomo@amazon.com>