[nvidia] On Ubuntu22, install the NVIDIA driver using the gcc version used to compile the kernel.
This is required because, NVIDIA driver must be compiled with the same gcc version used by the kernel.
If this is not the case, the NVIDIA driver installation would fail a compiler version check.
On newer version of Ubuntu22.04 (kernel 6.8+), the kernel is compiled with gcc-12, however gcc-11 is installed as default version by build-essentials, making this change necessary.

Signed-off-by: Giacomo Marciani <mgiacomo@amazon.com>